# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenEmbedded API — visual Discord embed builder
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Last compiled Discord payload
 * @nullable
 */
export type ProjectPayload = { [key: string]: unknown } | null;

export type GraphDataNodesItem = { [key: string]: unknown };

export type GraphDataEdgesItem = { [key: string]: unknown };

export interface GraphData {
  nodes: GraphDataNodesItem[];
  edges: GraphDataEdgesItem[];
}

export interface Project {
  id: string;
  name: string;
  graph: GraphData;
  /**
     * Last compiled Discord payload
     * @nullable
     */
  payload?: ProjectPayload;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  graph: GraphData;
}

/**
 * @nullable
 */
export type ProjectUpdatePayload = { [key: string]: unknown } | null;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  graph?: GraphData;
  /** @nullable */
  payload?: ProjectUpdatePayload;
}

export interface GraphInput {
  graph: GraphData;
}

export type ExportResultPayload = { [key: string]: unknown };

export interface ExportResult {
  payload: ExportResultPayload;
  isValid: boolean;
  errors?: string[];
}

export interface CodeExportResult {
  code: string;
}

export type WebhookSendInputPayload = { [key: string]: unknown };

export interface WebhookSendInput {
  webhookUrl: string;
  payload: WebhookSendInputPayload;
}

export interface WebhookSendResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface ErrorResponse {
  error: string;
}