# lib/api-zod/src/generated/types/exportResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenEmbedded API — visual Discord embed builder
 * OpenAPI spec version: 0.1.0
 */
import type { ExportResultPayload } from './exportResultPayload';

export interface ExportResult {
  payload: ExportResultPayload;
  isValid: boolean;
  errors?: string[];
}