# lib/api-zod/src/generated/types/graphData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenEmbedded API — visual Discord embed builder
 * OpenAPI spec version: 0.1.0
 */
import type { GraphDataEdgesItem } from './graphDataEdgesItem';
import type { GraphDataNodesItem } from './graphDataNodesItem';

export interface GraphData {
  nodes: GraphDataNodesItem[];
  edges: GraphDataEdgesItem[];
}