# lib/api-zod/src/generated/types/graphInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenEmbedded API — visual Discord embed builder
 * OpenAPI spec version: 0.1.0
 */
import type { GraphData } from './graphData';

export interface GraphInput {
  graph: GraphData;
}