# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenEmbedded API — visual Discord embed builder
 * OpenAPI spec version: 0.1.0
 */
import type { GraphData } from './graphData';
import type { ProjectUpdatePayload } from './projectUpdatePayload';

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  graph?: GraphData;
  /** @nullable */
  payload?: ProjectUpdatePayload;
}