# lib/api-zod/src/generated/types/webhookSendResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenEmbedded API — visual Discord embed builder
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookSendResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}